# command_center/openapi.json
{
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "answer": {
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          }
        },
        "required": [
          "answer"
        ],
        "title": "AnswerRequest",
        "type": "object"
      },
...
      "ArtifactContentOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "content",
          "truncated"
        ],
        "title": "ArtifactContentOut",
        "type": "object"
      },
...
      "BlockRequest": {
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "BlockRequest",
        "type": "object"
      },
...
      "DecisionRequest": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "pr",
              "artifact",
              "decision"
            ],
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "verdict": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "verdict"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "Finding": {
        "properties": {
          "claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "fix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "where": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Where"
          },
          "why": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why"
          }
        },
        "title": "Finding",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "merged": {
            "title": "Merged",
            "type": "boolean"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          }
        },
        "required": [
          "merged",
          "detail"
        ],
        "title": "MergeOut",
        "type": "object"
      },
      "MessageOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "PrDetailOut": {
        "properties": {
          "pr": {
            "$ref": "#/components/schemas/PrOut"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewOut"
            },
            "title": "Reviews",
            "type": "array"
          }
        },
        "required": [
          "pr",
          "reviews"
        ],
        "title": "PrDetailOut",
        "type": "object"
      },
      "PrOut": {
        "properties": {
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "em_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Em Summary"
          },
          "github_pr": {
            "title": "Github Pr",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "tyler_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tyler Decision"
          }
        },
        "required": [
          "id",
          "project_id",
          "github_pr",
          "status"
        ],
        "title": "PrOut",
        "type": "object"
      },
...
      "ReorderRequest": {
        "properties": {
          "ordered_ids": {
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Ordered Ids",
            "type": "array"
          }
        },
        "required": [
          "ordered_ids"
        ],
        "title": "ReorderRequest",
        "type": "object"
      },
...
      "ReviewOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "pr_id": {
            "title": "Pr Id",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "strongest_objection": {
            "title": "Strongest Objection",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pr_id",
          "role",
          "verdict",
          "strongest_objection",
          "created_at"
        ],
        "title": "ReviewOut",
        "type": "object"
      },
...
      "TaskOut": {
        "properties": {
          "blocked": {
            "title": "Blocked",
            "type": "boolean"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "depends_on": {
            "items": {
              "type": "integer"
            },
            "title": "Depends On",
            "type": "array"
          },
          "groom_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groom Rank"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "model_tier": {
            "title": "Model Tier",
            "type": "string"
          },
          "prd_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prd Ref"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "spec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "title",
          "status",
          "blocked",
          "model_tier",
          "depends_on",
          "created_at",
          "updated_at"
        ],
        "title": "TaskOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "emctl-over-HTTP: the platform state store given an authenticated JSON API for the command-center SPA.",
    "title": "Command Center API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/approvals/artifact/{artifact_id}/content": {
      "get": {
        "operationId": "artifact_content_api_approvals_artifact__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactContentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact Content",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/decision": {
      "post": {
        "operationId": "decide_api_approvals_decision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/pr/{pr_id}": {
      "get": {
        "operationId": "pr_detail_api_approvals_pr__pr_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pr_id",
            "required": true,
            "schema": {
              "title": "Pr Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pr Detail",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/pr/{pr_id}/merge": {
      "post": {
        "operationId": "merge_pr_api_approvals_pr__pr_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "pr_id",
            "required": true,
            "schema": {
              "title": "Pr Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Pr",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/question/{question_id}/answer": {
      "post": {
        "operationId": "answer_question_api_approvals_question__question_id__answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Question",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/backlog/reorder": {
      "post": {
        "operationId": "reorder_api_backlog_reorder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskOut"
                  },
                  "title": "Response Reorder Api Backlog Reorder Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder",
        "tags": [
          "backlog"
        ]
      }
    },
...
    "/api/tasks/{task_id}/block": {
      "post": {
        "operationId": "block_api_tasks__task_id__block_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/tasks/{task_id}/greenlight": {
      "post": {
        "operationId": "greenlight_api_tasks__task_id__greenlight_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Greenlight",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/tasks/{task_id}/unblock": {
      "post": {
        "operationId": "unblock_api_tasks__task_id__unblock_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unblock",
        "tags": [
          "backlog"
        ]
      }
    }
  }
}